modified the isweep package
