--- src/isweep/cis/__init__.py
@@ -1,18 +1,0 @@
-# packages
-import sys
-import os
-import gzip
-import numpy as np
-import pandas as pd
-import matplotlib.pyplot as plt
-import seaborn as sns
-from scipy.optimize import minimize, minimize_scalar
-from scipy.stats import binom, norm
-from math import sqrt, floor, ceil, exp, log
-from random import randint
-from copy import deepcopy
-
-# my modules
-from .coalescentIBD import *
-from .cisUtilities import *
-from .illustrative import *
