+++ src/isweep/favoredalleles.py
@@ -1,11 +1,7 @@
-#!/bin/python
-
 # imports
 import gzip
 import numpy as np
 import pandas as pd
-import matplotlib.pyplot as plt
-import seaborn as sns
 from copy import deepcopy
 
 ### relabeling, recoding ###
