+++ src/isweep/outgroups.py
@@ -1,5 +1,3 @@
-#!/bin/python
-
 # imports
 import numpy as np
 import networkx as nx
