+++ src/isweep/slow.py
